Ims Client Lead APIs modified

+++ ims-client/src/main/java/com/stanzaliving/core/ims/client/dto/responseDto/CityResponseDto.java
@@ -23,7 +23,7 @@ public class CityResponseDto {
 
    private String code;
 
-   private Long stateId;
+   private int stateId;
 
    private String stateName;
